FFApp updated
When FFApp generating colorbar as video frames to initialize FFE, the value of frame count has significant role. It should be one of the values from a predefined set. So, the frame count needs to be any value from the set of 80fps, 120fps, and 240fps. This is not applicable when FFApp initialize frames from a file.